refactor: Separate game logic from DOM methods
Major rewrite of the app to disentangle the 3 star
plate of spaghetti bolognese that was the main
driver module. Yum.

Now we have much better separation of concerns between
the DOM and the game logic. What's more, the game logic
can now be tested independently without having to play
a whole game in the browser.

This commit should've probably been split into several
for clarity, but the problem ran so deep that anything
but a whole rewrite would've broken the app. I don't
like to commit broken stuff. Only Linus can judge me now.